@@ -2,77 +2,32 @@
   "name": "admin",
   "version": "0.1.0",
   "private": true,
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "version": "echo $npm_package_version",
-    "postinstall": "is-ci || is-docker || husky install",
-    "prepublishOnly": "pinst --disable",
-    "postpublish": "pinst --enable",
-    "prepare": "husky install",
-    "lint": "eslint ."
-  },
   "dependencies": {
-    "@accounts/client": "^0.33.1",
-    "@accounts/client-password": "^0.32.2",
-    "@accounts/graphql-client": "^0.33.1",
-    "@apollo/client": "^3.3.15",
-    "@axa-fr/react-oidc-context": "^3.1.6",
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@mui/material": "^5.0.3",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "apollo-cache-inmemory": "^1.6.6",
-    "apollo-client": "^2.6.10",
-    "apollo-link": "^1.2.14",
-    "apollo-link-http": "^1.5.17",
-    "commitlint": "^12.0.1",
-    "dotenv": "^8.6.0",
-    "envalid": "^7.1.0",
-    "graphql": "^15.5.0",
-    "graphql-tag": "^2.12.0",
-    "http": "0.0.1-security",
-    "mdi-material-ui": "^6.22.1",
-    "oidc-client": "^1.11.5",
-    "ra-core": "^3.15.1",
-    "ra-data-graphql-simple": "^3.13.2",
-    "ra-data-json-server": "^3.13.4",
-    "react": "^17.0.1",
-    "react-admin": "^3.13.4",
-    "react-dom": "^17.0.1",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.2.4",
-    "url-parse": "^1.5.1",
-    "wait-on": "^5.3.0",
-    "web-vitals": "^1.0.1"
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.60",
+    "@types/react": "^18.2.34",
+    "@types/react-dom": "^18.2.14",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "recompose": "^0.30.0",
+    "typescript": "^4.9.5",
+    "web-vitals": "^2.1.4"
   },
-  "devDependencies": {
-    "@commitlint/cli": "^13.2.1",
-    "@commitlint/config-conventional": "^13.2.0",
-    "@reactioncommerce/eslint-config": "^2.2.0",
-    "@types/url-parse": "^1.4.3",
-    "@types/wait-on": "^5.2.0",
-    "@typescript-eslint/eslint-plugin": "^4.18.0",
-    "@typescript-eslint/parser": "^4.18.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.22.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.2.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.3.1",
-    "eslint-plugin-react": "^7.22.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "husky": "^7.0.2",
-    "ts-node-dev": "^1.1.6"
+  "scripts": {
+    "start": "craco start",
+    "build": "craco build",
+    "test": "craco test",
+    "eject": "craco eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -85,5 +40,10 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@commitlint/cli": "^18.4.3",
+    "@commitlint/config-conventional": "^18.4.3",
+    "@craco/craco": "^7.1.0"
   }
 }
